WIP - updated application init and file structure to match engine
