+++ services/organization/config/routes.rb
@@ -1,3 +1,5 @@
-Rails.application.routes.draw do
+# frozen_string_literal: true
+
+Ros::Organization::Engine.routes.draw do
   jsonapi_resources :orgs
 end
