--- services/organization/lib/organization.rb
@@ -1,6 +1,0 @@
-require 'ros/core'
-require "organization/engine"
-
-module Organization
-  # Your code goes here...
-end
