+++ services/organization/lib/ros/organization.rb
@@ -1,0 +1,8 @@
+require 'ros/core'
+require 'ros/organization/engine'
+
+module Ros
+  module Organization
+    # Your code goes here...
+  end
+end
