--- services/organization/spec/dummy/app/assets/config/manifest.js
@@ -1,1 +1,0 @@
-//= link_directory ../stylesheets .css
