+++ services/organization/spec/dummy/config/application.rb
@@ -1,27 +1,29 @@
+# frozen_string_literal: true
+
 require_relative 'boot'
 require 'rails-html-sanitizer'
-require "rails"
+require 'rails'
 # Pick the frameworks you want:
-require "active_model/railtie"
-require "active_job/railtie"
-require "active_record/railtie"
-# require "active_storage/engine"
-require "action_controller/railtie"
-# require "action_mailer/railtie"
-require "action_view/railtie"
-# require "action_cable/engine"
-# require "sprockets/railtie"
-# require "rails/test_unit/railtie"
+require 'active_model/railtie'
+require 'active_job/railtie'
+require 'active_record/railtie'
+# require 'active_storage/engine'
+require 'action_controller/railtie'
+# require 'action_mailer/railtie'
+require 'action_view/railtie'
+# require 'action_cable/engine'
+# require 'sprockets/railtie'
+# require 'rails/test_unit/railtie'
 
 Bundler.require(*Rails.groups)
-require "organization"
+require 'ros/organization'
 
 module Dummy
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 6.0
     config.hosts << 'organization'
-    config.hosts << 'localhost'
+    # config.hosts << 'localhost'
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration can go into files in config/initializers
@@ -34,4 +36,3 @@ module Dummy
     config.api_only = true
   end
 end
-
